Move prepare data submission to separate file
